requirements.txt
cachetools==5.5.2 ; python_version >= "3.12"
certifi==2025.4.26 ; python_version >= "3.12"
charset-normalizer==3.4.2 ; python_version >= "3.12"
google-api-core==2.25.0 ; python_version >= "3.12"
google-api-core[grpc]==2.25.0 ; python_version >= "3.12"
google-auth==2.40.3 ; python_version >= "3.12"
google-cloud-bigquery==3.34.0 ; python_version >= "3.12"
google-cloud-core==2.4.3 ; python_version >= "3.12"
google-cloud-storage==3.1.0 ; python_version >= "3.12"
google-crc32c==1.7.1 ; python_version >= "3.12"
google-resumable-media==2.7.2 ; python_version >= "3.12"
googleapis-common-protos==1.70.0 ; python_version >= "3.12"
grpcio==1.72.1 ; python_version >= "3.12"
grpcio-status==1.72.1 ; python_version >= "3.12"
idna==3.10 ; python_version >= "3.12"
memory-profiler==0.61.0 ; python_version >= "3.12"
numpy==2.3.0 ; python_version >= "3.12"
packaging==25.0 ; python_version >= "3.12"
pandas==2.3.0 ; python_version >= "3.12"
proto-plus==1.26.1 ; python_version >= "3.12"
protobuf==6.31.1 ; python_version >= "3.12"
psutil==7.0.0 ; python_version >= "3.12"
pyasn1==0.6.1 ; python_version >= "3.12"
pyasn1-modules==0.4.2 ; python_version >= "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.12"
python-dotenv==1.1.0 ; python_version >= "3.12"
pytz==2025.2 ; python_version >= "3.12"
requests==2.32.3 ; python_version >= "3.12"
rsa==4.2 ; python_version >= "4"
rsa==4.9.1 ; python_version >= "3.12" and python_version < "4"
six==1.17.0 ; python_version >= "3.12"
tzdata==2025.2 ; python_version >= "3.12"
urllib3==2.4.0 ; python_version >= "3.12"
